Wasmtime: Move `OnDemandInstanceAllocator` to its own module
Just code motion, no other changes.
